fix: reducers are DRY and have admin-specific thunks
