+++ client/store/cart.js
@@ -1,6 +1,5 @@
 import axios from 'axios';
 import { addOrder, updateOrder } from './orders';
-import { add as addError } from './error';
 
 /* -----------------    ACTION TYPES ------------------ */
 
@@ -71,10 +70,6 @@ export const fetchCart = user => (dispatch) => {
         dispatch(fetch(res.data[0]))
         dispatch(cartId(res.data[0].id))
       }
-    })
-    .catch((err) => {
-      dispatch(addError(err))
-      console.error('Unsucessful')
     });
 };
