+++ client/store/index.js
@@ -8,12 +8,9 @@ import products from './products';
 import reviews from './reviews';
 import { cart, cartIDred as cartId } from './cart';
 import error from './error';
-import adminOrders from './admin/orders';
-import adminProducts from './admin/products';
-import adminUsers from './admin/users';
 
 const reducer = combineReducers({
-  user, orders, products, cart, reviews, cartId, sessionId, error, adminOrders, adminProducts, adminUsers,
+  user, orders, products, cart, reviews, cartId, sessionId, error,
 });
 
 const middleware = composeWithDevTools(applyMiddleware(
@@ -32,7 +29,3 @@ export * from './products';
 export * from './reviews';
 export * from './error';
 export * from './cart';
-
-export * from './admin/orders';
-export * from './admin/products';
-export * from './admin/users';
