+++ client/store/orders.js
@@ -8,12 +8,16 @@ const FETCH_ORDERS = 'FETCH_ORDERS';
 const CREATE_ORDER = 'CREATE_ORDER';
 const REMOVE_ORDER = 'REMOVE_ORDER';
 const UPDATE_ORDER = 'UPDATE_ORDER';
+const ADMIN_UPDATE_ORDER = 'ADMIN_UPDATE_ORDER';
+
+
 /* ------------   ACTION CREATORS     ------------------ */
 
 const fetch = orders => ({ type: FETCH_ORDERS, orders });
 const create = order => ({ type: CREATE_ORDER, order });
 const remove = id => ({ type: REMOVE_ORDER, id });
 const update = order => ({ type: UPDATE_ORDER, order });
+const adminUpdate = order => ({ type: ADMIN_UPDATE_ORDER, order });
 
 
 /* ------------       REDUCER     ------------------ */
@@ -32,6 +36,11 @@ export default (orders = [], action) => {
       return orders.filter(order => order.id !== action.id);
 
     case UPDATE_ORDER:
+      return orders.map(order => (
+        action.order.id === order.id ? action.order : order
+      ));
+
+    case ADMIN_UPDATE_ORDER:
       return orders.map(order => (
         action.order.id === order.id ? action.order : order
       ));
@@ -90,5 +99,19 @@ export const purchaseOrder = (id, shipping, order) => (dispatch) => {
     .catch((err) => {
       dispatch(addError(err.response.statusText))
       console.error(`Updating order: ${shipping} unsuccesful`, err)
+    });
+};
+
+
+// ADMIN THUNKS
+
+export const adminUpdateOrder = (id, order) => (dispatch) => {
+  axios.put(`/api/admin/orders/${id}`, order)
+    .then((res) => {
+      dispatch(adminUpdate(res.data))
+    })
+    .catch((err) => {
+      dispatch(addError(err.response.statusText))
+      console.error(`Updating order: ${order} unsuccesful`, err)
     });
 };
