+++ server/api/admin/products.js
@@ -3,18 +3,6 @@ const { Product, User } = require('../../db/models');
 
 module.exports = router
 
-router.get('/', (req, res, next) => {
-  Product.findAll({ include: [{ model: User, attributes: ['id', 'email'], nested: true }] })
-    .then(product => res.json(product))
-    .catch(next);
-});
-
-router.get('/:id', (req, res, next) => {
-  Product.findById(req.params.id, { include: [{ model: User, attributes: ['id', 'email'], nested: true }] })
-    .then(product => res.json(product))
-    .catch(next);
-});
-
 router.post('/', (req, res, next) => {
   Product.create(req.body)
     .then(product => res.status(201).json(product))
